+++ src/error.rs
@@ -6,6 +6,7 @@ pub enum Error {
     WrongTag,
     UnknownCharacter,
     Overflow,
+    NoPadding,
 }
 
 impl fmt::Display for Error {
@@ -14,6 +15,7 @@ impl fmt::Display for Error {
             Self::WrongTag => f.write_str("id has an invalid tag"),
             Self::UnknownCharacter => f.write_str("id contains invalid characters"),
             Self::Overflow => f.write_str("id is too long"),
+            Self::NoPadding => f.write_str("id is in an unexpected format"),
         }
     }
 }